WASM CI: enforce the co-load probe + fix the h5py-interop test dtype
Now that the co-load fix works, drop the `|| echo` guards so coload_repro.py and
the full suite are enforced regression tests: any wasm co-load regression fails
the job.

Also fix testScalarHdf5Fields, which ran for the first time on wasm (h5py is now
present in the test env): on wasm32 h5py writes the scalar with its native int
(int32) while np.array([45]) defaults to int64, so the store type-mismatched the
component ("store from Python array of type 'LONG' into Record Component of type
'INT'"). Store the component's own dtype instead of assuming the two agree.

+++ test/python/unittest/API/APITest.py
@@ -2333,7 +2333,10 @@ class APITest(unittest.TestCase):
 
         series_read_write = io.Series(file, io.Access.read_write)
         E_x = series_read_write.iterations[0].meshes["E"]["x"]
-        E_x[:] = np.array([45])
+        # match the component's own dtype: h5py wrote the scalar with its native
+        # int (int32 on wasm32), while np.array([45]) defaults to int64 there, so
+        # store the component's dtype rather than assume they agree.
+        E_x[:] = np.array([45], dtype=E_x.dtype)
         series_read_write.close()
 
         series_read_again = io.Series(file, io.Access.read_only)
